refactor: NewWorksForYou section
